refactor: ProblemDetailsMiddleware
